Added stylecop, documented, added repository interfaces
