+++ TechTest/AnyCompany.Tests/Class1.cs
@@ -1,11 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿// <copyright file="Class1.cs" company="Investec Bank">
+// Copyright © Investec Bank 2018
+// </copyright>
 
 namespace AnyCompany.Tests
 {
+    /// <summary>
+    /// A test class.
+    /// </summary>
     public class Class1
     {
     }
