--- TechTest/AnyCompany/Customer.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace AnyCompany
-{
-    public class Customer
-    {
-        public string Country { get; set; }
-
-        public DateTime DateOfBirth { get; set; }
-
-        public string Name { get; set; }
-    }
-}
