--- TechTest/AnyCompany/CustomerRepository.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using System.Data.SqlClient;
-
-namespace AnyCompany
-{
-    public static class CustomerRepository
-    {
-        private static string ConnectionString = @"Data Source=(local);Database=Customers;User Id=admin;Password=password;";
-
-        public static Customer Load(int customerId)
-        {
-            Customer customer = new Customer();
-
-            SqlConnection connection = new SqlConnection(ConnectionString);
-            connection.Open();
-
-            SqlCommand command = new SqlCommand("SELECT * FROM Customer WHERE CustomerId = " + customerId,
-                connection);
-            var reader = command.ExecuteReader();
-
-            while (reader.Read())
-            {
-                customer.Name = reader["Name"].ToString();
-                customer.DateOfBirth = DateTime.Parse(reader["DateOfBirth"].ToString());
-                customer.Country = reader["Country"].ToString();
-            }
-
-            connection.Close();
-
-            return customer;
-        }
-    }
-}
