--- TechTest/AnyCompany/Order.cs
@@ -1,9 +1,0 @@
-﻿namespace AnyCompany
-{
-    public class Order
-    {
-        public int OrderId { get; set; }
-        public double Amount { get; set; }
-        public double VAT { get; set; }
-    }
-}
