--- TechTest/AnyCompany/OrderRepository.cs
@@ -1,25 +1,0 @@
-﻿using System.Data.SqlClient;
-
-namespace AnyCompany
-{
-    internal class OrderRepository
-    {
-        private static string ConnectionString = @"Data Source=(local);Database=Orders;User Id=admin;Password=password;";
-
-        public void Save(Order order)
-        {
-            SqlConnection connection = new SqlConnection(ConnectionString);
-            connection.Open();
-
-            SqlCommand command = new SqlCommand("INSERT INTO Orders VALUES (@OrderId, @Amount, @VAT)", connection);
-
-            command.Parameters.AddWithValue("@OrderId", order.OrderId);
-            command.Parameters.AddWithValue("@Amount", order.Amount);
-            command.Parameters.AddWithValue("@VAT", order.VAT);
-
-            command.ExecuteNonQuery();
-
-            connection.Close();
-        }
-    }
-}
