+++ TechTest/AnyCompany/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="StyleCop.Analyzers" version="1.1.118" targetFramework="net461" developmentDependency="true" />
+</packages>
